rename main screen
