--- lib/first_screen.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-
-class FirstScreen extends StatefulWidget {
-  const FirstScreen({super.key});
-
-  @override
-  State<FirstScreen> createState() => _FirstScreen();
-}
-
-class _FirstScreen extends State<FirstScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: const BoxDecoration(
-        color: Color.fromARGB(255, 230, 189, 237),
-      ),
-    );
-  }
-}
