+++ lib/main.dart
@@ -1,86 +1,18 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
+import 'package:ponder_points_app/quotes_screen.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  MyApp({super.key});
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       title: 'Flutter Demo',
-      home: const MyHomePage(),
+      home: QuotesScreen(),
     );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({
-    super.key,
-  });
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  List<String> quotes = [
-    '"The only way to achieve the impossible is to believe it is possible."',
-    '"Believe you can and you are halfway there."',
-    '"The only limit to our realization of tomorrow will be our doubts of today." ',
-    '"The future belongs to those who believe in the beauty of their dreams."',
-    '"Dont be afraid to give up the good to go for the great."',
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Color.fromARGB(255, 231, 181, 243),
-      appBar: AppBar(
-        backgroundColor: Color.fromARGB(255, 231, 181, 243),
-        elevation: 0,
-        centerTitle: true,
-        title: const Text(
-          'Ponder Points',
-        ),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Padding(
-              padding: const EdgeInsets.all(20),
-              child: Text(
-                getQuote(),
-                textAlign: TextAlign.center,
-                style: TextStyle(color: Colors.white, fontSize: 30),
-              ),
-            ),
-            SizedBox(
-              width: 40,
-              height: 40,
-            ),
-            IconButton(
-              onPressed: () {},
-              icon: Icon(Icons.favorite_border),
-              iconSize: 50,
-              color: Colors.pink,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  String getQuote() {
-    final _random = Random();
-    final randomInt = _random.nextInt(quotes.length);
-    print(randomInt);
-    var quote = quotes[randomInt];
-    return quote;
   }
 }
